+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 * Override val not respected. [#38](https://github.com/nojaf/telplin/issues/38)
+* Constraint is missing from binding. [#39](https://github.com/nojaf/telplin/issues/39)
 
 ## 0.3.1 - 25-04-22